refactor: DiningPeriod 클래스를 NotificationRequest 도메인의 하위 패키지로 이동
