+++ src/main/java/com/jvnlee/catchdining/domain/notification/dto/NotificationRequestViewDto.java
@@ -1,7 +1,7 @@
 package com.jvnlee.catchdining.domain.notification.dto;
 
 import com.jvnlee.catchdining.domain.notification.model.NotificationRequest;
-import com.jvnlee.catchdining.entity.DiningPeriod;
+import com.jvnlee.catchdining.domain.notification.model.DiningPeriod;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
