+++ src/main/java/com/jvnlee/catchdining/domain/notification/model/DiningPeriod.java
@@ -1,0 +1,5 @@
+package com.jvnlee.catchdining.domain.notification.model;
+
+public enum DiningPeriod {
+    LUNCH, DINNER
+}
