--- src/main/java/com/jvnlee/catchdining/entity/DiningPeriod.java
@@ -1,5 +1,0 @@
-package com.jvnlee.catchdining.entity;
-
-public enum DiningPeriod {
-    LUNCH, DINNER
-}
